Add syntax highlighting
Using Prism with Oceanic Next, punched up with a filter. I may try and
move the override into a custom theme, but this works for now.

I had to add a utility style thanks to `pre`'s unformatted whitespace,
which causes grid columns, which have a min of `auto` to not collapse
like I want them to. The fix is `min-width: 0`, which is easily done via
a utility class.

I am also adding a passthrough for JS. We'll eventually tool up the JS
workflow as needed.