Adjust project scripts
After doing some digging, I found out that the ampersand pushes a task
to the background, allowing another task to be run on top of it. The
problem with this is there's no easy way of knowing when the background
task finishes, so if the foreground task finishes before, Netlify will
throw an error. npm-run-all fixes this, so I can go back to the standard
build script.